TMS_32 1)Created additional user dto for updating 2)Improved user Service and Controller - added additional checks 3)Updated ddl 4)Created pre-deploy script - for users, barbers, schedule, hair services creation (sql script and procedure)

+++ TMS_Booking_Project/src/main/java/com/teachmeskills/tms_booking_project/service/UserService.java
@@ -5,14 +5,18 @@ import com.teachmeskills.tms_booking_project.model.User;
 import com.teachmeskills.tms_booking_project.model.dto.UserCreateRequest;
 import com.teachmeskills.tms_booking_project.model.dto.UserRegistrationRequest;
 import com.teachmeskills.tms_booking_project.model.dto.UserResponse;
+import com.teachmeskills.tms_booking_project.model.dto.UserUpdateRequest;
 import com.teachmeskills.tms_booking_project.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -21,6 +25,7 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private static final Logger logger = LoggerFactory.getLogger(UserService.class);
 
     @Transactional
     public UserResponse register(UserRegistrationRequest request) {
@@ -32,11 +37,11 @@ public class UserService {
                 .fullName(request.fullName())
                 .email(request.email())
                 .password(passwordEncoder.encode(request.password()))
-                .role(Role.USER.name())
+                .role(Role.USER)
                 .registeredAt(LocalDateTime.now())
                 .isSubscribed(false)
                 .build();
-
+        logger.info("Registered user with id: {}", user.getId());
         return mapToResponse(userRepository.save(user));
     }
 
@@ -50,11 +55,11 @@ public class UserService {
                 .fullName(request.fullName())
                 .email(request.email())
                 .password(passwordEncoder.encode(request.password()))
-                .role(request.role().toUpperCase())
+                .role(request.role())
                 .isSubscribed(request.isSubscribed())
                 .registeredAt(LocalDateTime.now())
                 .build();
-
+        logger.info("Created user with id: {}", user.getId());
         return mapToResponse(userRepository.save(user));
     }
 
@@ -71,25 +76,59 @@ public class UserService {
     }
 
     @Transactional
-    public UserResponse updateUser(Long id, UserCreateRequest request) {
+    public UserResponse updateUser(Long id, UserUpdateRequest request) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("User not found"));
 
-        user.setFullName(request.fullName());
-        user.setEmail(request.email());
-        user.setSubscribed(request.isSubscribed());
-        user.setRole(request.role().toUpperCase());
+        boolean hasChanged = false;
 
-        if (request.password() != null && !request.password().isBlank()) {
-            user.setPassword(passwordEncoder.encode(request.password()));
+        if (!Objects.equals(request.fullName(), user.getFullName()) && request.fullName() != null) {
+            user.setFullName(request.fullName());
+            hasChanged = true;
         }
 
+        if (!Objects.equals(request.email(), user.getEmail()) && request.email() != null) {
+            if (userRepository.existsByEmail(request.email())) {
+                throw new IllegalArgumentException("Email already in use");
+            }
+            user.setEmail(request.email());
+            hasChanged = true;
+        }
+
+        if (request.isSubscribed() != null && request.isSubscribed() != user.isSubscribed()) {
+            user.setSubscribed(request.isSubscribed());
+            hasChanged = true;
+        }
+
+        if (request.role() != user.getRole() && request.role() != null) {
+            user.setRole(request.role());
+            hasChanged = true;
+        }
+
+        if (request.password() != null && !request.password().isBlank()) {
+            if (!passwordEncoder.matches(request.password(), user.getPassword())) {
+                user.setPassword(passwordEncoder.encode(request.password()));
+                hasChanged = true;
+            }
+        }
+
+        if (!hasChanged) {
+            logger.info("No changes detected for user {}", id);
+            throw new IllegalArgumentException("No changes detected");
+        }
+
+        logger.info("Updated user {} - changes applied", id);
         return mapToResponse(userRepository.save(user));
     }
 
     public boolean deleteUser(Long id) {
-        userRepository.deleteById(id);
-        return true;
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            logger.info("Removed user with id: {}", id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public UserResponse mapToResponse(User user) {
